+++ cvat-ui/src/reducers/annotation-reducer.ts
@@ -287,7 +287,7 @@ export default (state = defaultState, action: AnyAction): AnnotationState => {
                     },
                     contextImage: {
                         ...state.player.contextImage,
-                        data: null,
+                        ...(state.player.frame.number === number ? {} : { data: null }),
                     },
                 },
                 annotations: {